+++ realm/realm-library/src/androidTest/java/io/realm/TypeBasedNotificationsTests.java
@@ -24,7 +24,6 @@ import android.util.Base64;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -249,21 +248,6 @@ public class TypeBasedNotificationsTests {
         assumeThat(Build.VERSION.SDK_INT, greaterThanOrEqualTo(Build.VERSION_CODES.HONEYCOMB));
 
         final Realm realm = looperThread.realm;
-        realm.addChangeListener(new RealmChangeListener<Realm>() {
-            @Override
-            public void onChange(Realm object) {
-                if (globalCommitInvocations.incrementAndGet() == 1) {
-                    looperThread.postRunnable(new Runnable() {
-                        @Override
-                        public void run() {
-                            assertEquals(1, typebasedCommitInvocations.get());
-                            looperThread.testComplete();
-                        }
-                    });
-                }
-            }
-        });
-
         try {
             InputStream in = TestHelper.loadJsonFromAssets(InstrumentationRegistry.getTargetContext(), "all_simple_types.json");
             realm.beginTransaction();
@@ -281,7 +265,7 @@ public class TypeBasedNotificationsTests {
                     assertEquals(1.23D, objectFromJson.getColumnDouble(), 0D);
                     assertEquals(true, objectFromJson.isColumnBoolean());
                     assertArrayEquals(new byte[]{1, 2, 3}, objectFromJson.getColumnBinary());
-                    typebasedCommitInvocations.incrementAndGet();
+                    looperThread.testComplete();
                 }
             });
 
@@ -299,20 +283,6 @@ public class TypeBasedNotificationsTests {
     @RunTestInLooperThread
     public void callback_should_trigger_for_createObjectFromJson_from_JSONObject() {
         final Realm realm = looperThread.realm;
-        realm.addChangeListener(new RealmChangeListener<Realm>() {
-            @Override
-            public void onChange(Realm object) {
-                if (globalCommitInvocations.incrementAndGet() == 1) {
-                    looperThread.postRunnable(new Runnable() {
-                        @Override
-                        public void run() {
-                            assertEquals(1, typebasedCommitInvocations.get());
-                            looperThread.testComplete();
-                        }
-                    });
-                }
-            }
-        });
 
         try {
             JSONObject json = new JSONObject();
@@ -337,7 +307,7 @@ public class TypeBasedNotificationsTests {
                     assertEquals(1.23D, objectFromJson.getColumnDouble(), 0D);
                     assertEquals(true, objectFromJson.isColumnBoolean());
                     assertArrayEquals(new byte[]{1, 2, 3}, objectFromJson.getColumnBinary());
-                    typebasedCommitInvocations.incrementAndGet();
+                    looperThread.testComplete();
                 }
             });
 
@@ -708,20 +678,6 @@ public class TypeBasedNotificationsTests {
     public void multiple_callbacks_should_be_invoked_realmobject_async() {
         final int NUMBER_OF_LISTENERS = 7;
         final Realm realm = looperThread.realm;
-        RealmChangeListener<Realm> listener = new RealmChangeListener<Realm>() {
-            @Override
-            public void onChange(Realm object) {
-                looperThread.postRunnable(new Runnable() {
-                    @Override
-                    public void run() {
-                        assertEquals(NUMBER_OF_LISTENERS, typebasedCommitInvocations.get());
-                        looperThread.testComplete();
-                    }
-                });
-            }
-        };
-
-        realm.addChangeListener(listener);
 
         realm.beginTransaction();
         Dog akamaru = realm.createObject(Dog.class);
@@ -735,6 +691,17 @@ public class TypeBasedNotificationsTests {
                 @Override
                 public void onChange(Dog object) {
                     typebasedCommitInvocations.incrementAndGet();
+                    if (typebasedCommitInvocations.get() > NUMBER_OF_LISTENERS) {
+                        fail();
+                    } else if (typebasedCommitInvocations.get() == NUMBER_OF_LISTENERS) {
+                        // Delayed post in case the listener gets triggered more time than expected.
+                        looperThread.postRunnableDelayed(new Runnable() {
+                            @Override
+                            public void run() {
+                                looperThread.testComplete();
+                            }
+                        }, 500);
+                    }
                 }
             });
         }
@@ -830,20 +797,6 @@ public class TypeBasedNotificationsTests {
     @RunTestInLooperThread
     public void non_looper_thread_commit_realmobject_sync() {
         final Realm realm = looperThread.realm;
-        realm.addChangeListener(new RealmChangeListener<Realm>() {
-            @Override
-            public void onChange(Realm object) {
-                if (realm.where(Dog.class).count() == 2) {
-                    looperThread.postRunnable(new Runnable() {
-                        @Override
-                        public void run() {
-                            assertEquals(1, typebasedCommitInvocations.get());
-                            looperThread.testComplete();
-                        }
-                    });
-                }
-            }
-        });
 
         realm.beginTransaction();
         realm.createObject(Dog.class);
@@ -854,73 +807,31 @@ public class TypeBasedNotificationsTests {
         dog.addChangeListener(new RealmChangeListener<Dog>() {
             @Override
             public void onChange(Dog object) {
-                typebasedCommitInvocations.incrementAndGet();
+                assertEquals(17, object.getAge());
+                looperThread.testComplete();
             }
         });
 
-        Thread thread = new Thread() {
+        realm.executeTransactionAsync(new Realm.Transaction() {
             @Override
-            public void run() {
-                Realm bgRealm = Realm.getInstance(realm.getConfiguration());
-                bgRealm.beginTransaction();
-                bgRealm.createObject(Dog.class);
-                bgRealm.commitTransaction();
-                bgRealm.close();
+            public void execute(Realm realm) {
+                realm.where(Dog.class).findFirst().setAge(17);
             }
-        };
-        thread.start();
-        try {
-            thread.join();
-        } catch (InterruptedException e) {
-            fail(e.getMessage());
-        }
+        });
     }
 
     // UC 3 Async RealmObject.
     // 1. Creates RealmObject async query.
-    // 2. Waits COMPLETED_ASYNC_REALM_OBJECT then commits transaction in another non-looper thread.
+    // 2. Waits async returns then change the object.
     // 3. Listener on the RealmObject gets triggered again.
     @Test
     @RunTestInLooperThread
     public void non_looper_thread_commit_realmobject_async() {
         final Realm realm = looperThread.realm;
-        realm.addChangeListener(new RealmChangeListener<Realm>() {
-            @Override
-            public void onChange(Realm object) {
-                // Checks if the 2nd transaction is committed.
-                if (realm.where(Dog.class).count() == 2) {
-                    looperThread.postRunnable(new Runnable() {
-                        @Override
-                        public void run() {
-                            assertEquals(2, typebasedCommitInvocations.get());
-                            looperThread.testComplete();
-                        }
-                    });
-                }
-            }
-        });
 
         realm.beginTransaction();
-        realm.createObject(Dog.class);
+        realm.createObject(Dog.class).setAge(0);
         realm.commitTransaction();
-
-        final Thread thread = new Thread() {
-            @Override
-            public void run() {
-                if (typebasedCommitInvocations.get() != 1) {
-                    try {
-                        Thread.sleep(200);
-                    } catch (InterruptedException e) {
-                        fail(e.getMessage());
-                    }
-                }
-                Realm bgRealm = Realm.getInstance(realm.getConfiguration());
-                bgRealm.beginTransaction();
-                bgRealm.createObject(Dog.class);
-                bgRealm.commitTransaction();
-                bgRealm.close();
-            }
-        };
 
         Dog dog = realm.where(Dog.class).findFirstAsync();
         looperThread.keepStrongReference.add(dog);
@@ -929,17 +840,25 @@ public class TypeBasedNotificationsTests {
             public void onChange(Dog object) {
                 typebasedCommitInvocations.incrementAndGet();
 
-                if (typebasedCommitInvocations.get() == 1) {
-                    try {
-                        thread.join();
-                    } catch (InterruptedException e) {
-                        fail(e.getMessage());
-                    }
+                switch (typebasedCommitInvocations.get()) {
+                    case 1:
+                        assertEquals(0, object.getAge());
+                        realm.executeTransactionAsync(new Realm.Transaction() {
+                            @Override
+                            public void execute(Realm realm) {
+                                realm.where(Dog.class).findFirst().setAge(17);
+                            }
+                        });
+                        break;
+                    case 2:
+                        assertEquals(17, object.getAge());
+                        looperThread.testComplete();
+                        break;
+                    default:
+                        fail();
                 }
             }
         });
-
-        thread.start();
     }
 
     // UC 3 Sync RealmResults.